Fixes issue 39
Fixes issue 39 https://github.com/UndefinedOffset/silverstripe-nocaptcha/issues/39
It's a scope issue caused by rendering the fields in a loop resulting in the last field rendered to always be the one processed when any form submits. This solution breaks the rendering into a separate function and then just call that function in the loop.